refactor(checkpoint): share passthrough adapter conversion
